# gears/system/resource-group/docs/openapi.yaml
...
openapi: 3.1.0
info:
  title: Resource Group API
  version: 0.1.0
  description: |
    Resource Group API — a generic hierarchy and membership engine.

    **This document is generated.** It is the resource-group subset of `docs/api/api.json`, which
    `make openapi` produces from the running server. Do not edit it by hand — edit
    `docs/openapi.preamble.yaml` (this preamble) or the code, then run `make openapi-gears`.

    ## Path prefix

    The paths below are gear-relative, exactly as the gear registers them. A deployment may nest
    them under a gateway prefix configured as `api-gateway.prefix_path`; in `config/quickstart.yaml`
    that prefix is `/cf`, and in the local e2e stack there is none. The `servers` entry reflects the
    configuration the document was generated from.

    ## Types API

    `/types-registry/v1/types` is a structured registry, not a raw schema store. A type is created
    from a flat DTO — `code`, `can_be_root`, `allowed_parent_types`, `allowed_membership_types`,
    `metadata_schema` — and is addressed by its `code` (a GTS type path). `can_be_root` has no
    column of its own: it is persisted inside `metadata_schema` under the reserved key
    `__can_be_root`, and reserved `__`-prefixed keys are stripped on read.

    Registration order matters: types referenced in `allowed_parent_types` or
    `allowed_membership_types` must already exist, so base types come before chained ones. Seeding
    is strictly sequential for this reason.

...

...

...

...

...

...

...

...

...
servers:
...
paths:
  /resource-group/v1/groups:
    get:
      description: Retrieve a paginated list of resource groups with OData filtering
      operationId: resource_group.list_groups
      parameters:
      - description: Maximum number of groups to return
        in: query
        name: limit
        required: false
        schema:
          type: integer
      - description: Cursor for pagination
        in: query
        name: cursor
        required: false
        schema:
          type: string
      - description: |-
          OData v4 filter expression
          - type: eq|ne|contains|startswith|endswith|in
          - hierarchy/parent_id: eq|ne|in
          - tenant_id: eq|ne|in
          - id: eq|ne|in
          - name: eq|ne|contains|startswith|endswith|in
        in: query
        name: $filter
        required: false
        schema:
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page_GroupDto'
          description: Paginated list of resource groups
        '400':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Bad Request
        '401':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Unauthorized
        '403':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Forbidden
        '500':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Internal Server Error
      security:
      - bearerAuth: []
      summary: List resource groups
      tags:
      - Resource Groups
      x-odata-filter:
        allowedFields:
          hierarchy/parent_id:
          - eq
          - ne
          - in
          id:
          - eq
          - ne
          - in
          name:
          - eq
          - ne
          - contains
          - startswith
          - endswith
          - in
          tenant_id:
          - eq
          - ne
          - in
          type:
          - eq
          - ne
          - contains
          - startswith
          - endswith
          - in
    post:
      description: Create a new resource group with the provided type, name, and optional parent
      operationId: resource_group.create_group
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGroupDto'
        description: Group creation data
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupDto'
          description: Created resource group
        '400':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Bad Request
        '401':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Unauthorized
        '403':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Forbidden
        '404':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Not Found
        '409':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Conflict
        '500':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Internal Server Error
      security:
      - bearerAuth: []
      summary: Create a new resource group
      tags:
      - Resource Groups
  /resource-group/v1/groups/{group_id}:
    delete:
      description: Delete a resource group. Use ?force=true to cascade delete subtree and memberships.
      operationId: resource_group.delete_group
      parameters:
      - description: Group UUID
        in: path
        name: group_id
        required: true
        schema:
          type: string
      - description: Force cascade delete of subtree and memberships
        in: query
        name: force
        required: false
        schema:
          type: boolean
      responses:
        '204':
          description: Group deleted successfully
        '400':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Bad Request
        '401':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Unauthorized
        '403':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Forbidden
        '404':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Not Found
        '500':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Internal Server Error
      security:
      - bearerAuth: []
      summary: Delete resource group
      tags:
      - Resource Groups
    get:
      description: Retrieve a specific resource group by its UUID
      operationId: resource_group.get_group
      parameters:
      - description: Group UUID
        in: path
        name: group_id
        required: true
        schema:
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupDto'
          description: Resource group found
        '400':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Bad Request
        '401':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Unauthorized
        '403':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Forbidden
        '404':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Not Found
        '500':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Internal Server Error
      security:
      - bearerAuth: []
      summary: Get resource group by ID
      tags:
      - Resource Groups
    put:
      description: 'Replace a resource group''s ordinary attributes: `name` and `metadata`. Both keys
        are required -- an omitted key is a 400, not "keep the stored value"; send `"metadata": null`
        to clear the metadata. The group''s `type` is immutable and its parent is not part of this payload:
        use POST /resource-group/v1/groups/{group_id}/move to re-parent a group. Unknown fields are rejected.'
      operationId: resource_group.update_group
      parameters:
      - description: Group UUID
        in: path
        name: group_id
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateGroupDto'
        description: Group update data
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupDto'
          description: Updated resource group
        '400':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Bad Request
        '401':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Unauthorized
        '403':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Forbidden
        '404':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Not Found
        '500':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Internal Server Error
      security:
      - bearerAuth: []
      summary: Update resource group
      tags:
      - Resource Groups
  /resource-group/v1/groups/{group_id}/ancestors:
    get:
      description: Get ancestors of a reference group (depth <= 0) with OData filtering
      operationId: resource_group.get_group_ancestors
      parameters:
      - description: Reference group UUID
        in: path
        name: group_id
        required: true
        schema:
          type: string
      - description: Maximum entries to return
        in: query
        name: limit
        required: false
        schema:
          type: integer
      - description: Cursor for pagination
        in: query
        name: cursor
        required: false
        schema:
          type: string
      - description: |-
          OData v4 filter expression
          - hierarchy/depth: eq|ne|gt|ge|lt|le
          - type: eq|ne|in
        in: query
        name: $filter
        required: false
        schema:
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page_GroupWithDepthDto'
          description: Paginated ancestors with relative depth
        '400':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Bad Request
        '401':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Unauthorized
        '403':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Forbidden
        '404':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Not Found
        '500':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Internal Server Error
      security:
      - bearerAuth: []
      summary: Get group ancestors
      tags:
      - Resource Groups
  /resource-group/v1/groups/{group_id}/descendants:
    get:
      description: Get descendants of a reference group (depth >= 0) with OData filtering
      operationId: resource_group.get_group_descendants
      parameters:
      - description: Reference group UUID
        in: path
        name: group_id
        required: true
        schema:
          type: string
      - description: Maximum entries to return
        in: query
        name: limit
        required: false
        schema:
          type: integer
      - description: Cursor for pagination
        in: query
        name: cursor
        required: false
        schema:
          type: string
      - description: |-
          OData v4 filter expression
          - hierarchy/depth: eq|ne|gt|ge|lt|le
          - type: eq|ne|in
        in: query
        name: $filter
        required: false
        schema:
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page_GroupWithDepthDto'
          description: Paginated descendants with relative depth
        '400':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Bad Request
        '401':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Unauthorized
        '403':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Forbidden
        '404':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Not Found
        '500':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Internal Server Error
      security:
      - bearerAuth: []
      summary: Get group descendants
      tags:
      - Resource Groups
  /resource-group/v1/groups/{group_id}/move:
    post:
      description: 'Atomically move a resource group -- and its entire subtree -- to a new parent, or
        to the forest root. `parent_id` is required: an explicit `null` means "make this group a root",
        while omitting the key is a 400. Cycle detection, parent-type compatibility, depth/width limits,
        tenant-root uniqueness and the closure-table rebuild all run inside one SERIALIZABLE transaction,
        so the group is never left detached from both parents. Cross-tenant moves are rejected.'
      operationId: resource_group.move_group
      parameters:
      - description: Group UUID
        in: path
        name: group_id
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MoveGroupDto'
        description: Move destination
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupDto'
          description: Moved resource group
        '400':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Bad Request
        '401':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Unauthorized
        '403':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Forbidden
        '404':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Not Found
        '409':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Conflict
        '500':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Internal Server Error
      security:
      - bearerAuth: []
      summary: Move resource group to a new parent
      tags:
      - Resource Groups
...
components:
  schemas:
    CreateGroupDto:
      description: REST DTO for creating a new resource group.
      properties:
        id:
          description: Optional caller-supplied ID. If omitted, the server generates a UUID.
          format: uuid
          type:
          - string
          - 'null'
        metadata:
          description: Type-specific metadata.
        name:
          description: Display name (1..255 characters).
          type: string
        parent_id:
          description: Parent group ID (null for root groups).
          format: uuid
          type:
          - string
          - 'null'
        tenant_id:
          description: |-
            Optional target tenant for the created group (VHP-2162).

            If omitted, the tenant scope is derived from the caller's own
            `SecurityContext` -- today's unchanged default behavior. If present
            and different from the caller's own tenant, the create succeeds only
            when the caller's `create`-action `AccessScope` actually covers the
            target tenant (platform-admin / onboarding use case); otherwise the
            request is rejected as though the target tenant did not exist.
            Rejected as a contradiction for tenant-typed groups: their effective
            tenant is always `group.id`.
          format: uuid
          type:
          - string
          - 'null'
        type:
          description: GTS chained type path. Must have prefix `gts.cf.core.rg.type.v1~`.
          type: string
      required:
      - type
      - name
      type: object
...
    GroupDto:
      description: |-
        REST DTO for resource group representation.

        Group responses do NOT include `created_at`/`updated_at` (per DESIGN).
      properties:
        hierarchy:
          $ref: '#/components/schemas/HierarchyDto'
          description: Hierarchy context.
        id:
          description: Group identifier.
          format: uuid
          type: string
        metadata:
          description: Type-specific metadata.
        name:
          description: Display name.
          type: string
        type:
          description: GTS chained type path.
          type: string
      required:
      - id
      - type
      - name
      - hierarchy
      type: object
...
    PageInfo:
      properties:
        limit:
          format: int64
          minimum: 0
          type: integer
        next_cursor:
          type:
          - string
          - 'null'
        prev_cursor:
          type:
          - string
          - 'null'
      required:
      - limit
      type: object
    Page_GroupDto:
      properties:
        items:
          items:
            $ref: '#/components/schemas/GroupDto'
          type: array
        page_info:
          $ref: '#/components/schemas/PageInfo'
      required:
      - items
      - page_info
      type: object
    Page_GroupWithDepthDto:
      properties:
        items:
          items:
            $ref: '#/components/schemas/GroupWithDepthDto'
          type: array
        page_info:
          $ref: '#/components/schemas/PageInfo'
      required:
      - items
      - page_info
      type: object
...
    Problem:
      description: RFC 9457 problem+json. `context` varies by error category.
      properties:
        context:
          type: object
        detail:
          type: string
        instance:
          type: string
        status:
          format: int32
          type: integer
        title:
          type: string
        trace_id:
          type: string
        type:
          type: string
      required:
      - type
      - title
      - status
      - detail
      - context
      type: object
...
  securitySchemes:
    bearerAuth:
      bearerFormat: JWT
      scheme: bearer
      type: http